+++ src/array/index.js
@@ -14,3 +14,4 @@ export { default as shortestStringLength } from './shortest-string-length.js';
 export { default as longestStringLength } from './longest-string-length.js';
 export { default as oddItems } from './odd-items.js';
 export { default as evenItems } from './even-items.js';
+export { default as mergeArrayUnique } from './merge-array-unique.js';