+++ src/array/merge-array-unique.js
@@ -1,0 +1,9 @@
+/**
+ * Returns a new Array which contains all the values of the two Arrays. Ensures that the new array contains only unique values.
+ * @param {Array} a 
+ * @param {Array} b 
+ * @returns {Number}
+ */
+const fn = (a, b) => [...new Set([...a, ...b])];
+
+export default fn;