+++ test/array/index.spec.js
@@ -62,3 +62,6 @@ test('exports a function "(Array.)oddItems"', t => {
 test('exports a function "(Array.)evenItems"', t => {
     t.true(typeof (Array.evenItems) === 'function')
 });
+test('exports a function "(Array.)mergeArrayUnique"', t => {
+    t.true(typeof (Array.mergeArrayUnique) === 'function')
+});